Refactor table, add error handling, fetching, etc
