+++ app/javascript/components/localization/originals/OriginalsTableListElement.tsx
@@ -1,0 +1,56 @@
+import React from 'react'
+import { GraphicalIcon } from '@/components/common'
+import { OriginalsListContext } from '.'
+import { assembleClassNames } from '@/utils/assemble-classnames'
+import { flagForLocale } from '@/utils/flag-for-locale'
+
+export function OriginalsTableListElement({
+  original,
+}: {
+  original: Original
+}) {
+  const { links } = React.useContext(OriginalsListContext)
+  return (
+    <a
+      href={links?.localizationOriginalsPath + '/' + original.uuid}
+      className="original"
+    >
+      <div className="info">
+        <div className="original-key">{original.key}</div>
+        <div className="original-uuid">{original.uuid}</div>
+      </div>
+
+      <TranslationsWithStatus translations={original.translations} />
+
+      <div className="rhs">
+        <div className="translation-glimpse">{original.value}</div>
+        <GraphicalIcon
+          icon="chevron-right"
+          className="action-icon filter-textColor6"
+        />
+      </div>
+    </a>
+  )
+}
+
+export function TranslationsWithStatus({
+  translations,
+}: {
+  translations: Original['translations']
+}) {
+  return (
+    <div className="translations-statuses">
+      {translations.map((translation) => (
+        <div
+          className={assembleClassNames(
+            'translation-status',
+            translation.status
+          )}
+          key={translation.uuid}
+        >
+          {flagForLocale(translation.locale)}
+        </div>
+      ))}
+    </div>
+  )
+}
